# key: optimize `Matches`