--- a/key/key_test.go
+++ b/key/key_test.go
@@ -2,6 +2,8 @@ package key
 
 import (
 	"testing"
+
+	tea "github.com/charmbracelet/bubbletea"
 )
 
 func TestBinding_Enabled(t *testing.T) {
@@ -23,4 +25,20 @@ func TestBinding_Enabled(t *testing.T) {
 	if binding.Enabled() {
 		t.Errorf("expected key not to be Enabled")
 	}
+}
+
+func BenchmarkMatches(b *testing.B) {
+	msg1 := tea.KeyMsg(tea.Key{Type: tea.KeyRunes, Runes: []rune("c"), Alt: true})
+	msg2 := tea.KeyMsg(tea.Key{Type: tea.KeyEnter})
+	kb := NewBinding(WithKeys("alt+c"))
+	b.Run("success", func(b *testing.B) {
+		for i := 0; i < b.N; i++ {
+			_ = Matches(msg1, kb)
+		}
+	})
+	b.Run("fail", func(b *testing.B) {
+		for i := 0; i < b.N; i++ {
+			_ = Matches(msg2, kb)
+		}
+	})
 }