@knz 2022-10-02
NB: I could imagine the map from string to tea.Key being defined in the bubbletea package directly. If you prefer that, let me knpw.