@knz 2022-10-06
> I think that makes sense and is probably nicer next to the other tea.Key code.

Ok see https://github.com/charmbracelet/bubbletea/pull/491

Then to merge this PR I'll need a release number for the go.mod update.